fix navbar style, fix table for progress change and make livetime update table
