+++ src/assets/css/index.css
@@ -4,11 +4,12 @@
   box-sizing: border-box;
 }
 
-html{
+body{
   background-image: url("../img/Tokopedia-Regist.jpg");
   background-position: center;
+  background-attachment: fixed;
+  background-size: cover;
 }
-
 
 #logo1 {  
     text-align: center;
@@ -238,7 +239,6 @@ select option {
 
 .container {
   overflow: hidden;
-  width: 700px;
   padding-left: 0;
 }
