+++ src/components/Header/Header.jsx
@@ -1,22 +1,45 @@
 import React, { Component } from 'react';
 import { Navbar } from 'react-bootstrap';
-
 import HeaderLinks from './HeaderLinks.jsx';
-
 import appRoutes from 'routes/app.jsx';
 
-class Header extends Component{
-    constructor(props){
+const $ = require('jquery');
+
+class Header extends Component {
+    constructor(props) {
         super(props);
         this.mobileSidebarToggle = this.mobileSidebarToggle.bind(this);
         this.state = {
             sidebarExists: false
         };
     }
-    mobileSidebarToggle(e){
-        if(this.state.sidebarExists === false){
+
+    componentDidMount() {
+        $(".navbar-container").css("width",this.sizeHandle());        
+        $(window).resize(function () {
+            let widthNavbar = ($(window).width()) - 258
+            if ($(window).width() >= 990) {
+                $(".navbar-container").width(widthNavbar);
+            }
+            else{
+                $(".navbar-container").css("width","100%");
+            }
+        });
+    }
+
+    sizeHandle(){
+        if ($(window).width() >= 990) {
+            return window.innerWidth - 260
+        }
+        else{
+            return "100%"
+        }
+    }
+
+    mobileSidebarToggle(e) {
+        if (this.state.sidebarExists === false) {
             this.setState({
-                sidebarExists : true
+                sidebarExists: true
             });
 
         }
@@ -24,29 +47,29 @@ class Header extends Component{
         document.documentElement.classList.toggle('nav-open');
         var node = document.createElement('div');
         node.id = 'bodyClick';
-        node.onclick = function(){
+        node.onclick = function () {
             this.parentElement.removeChild(this);
             document.documentElement.classList.toggle('nav-open');
         };
         document.body.appendChild(node);
     }
-    getBrand(){
+    getBrand() {
         var name;
-        appRoutes.map((prop,key) => {
-            if(prop.collapse){
-                 prop.views.map((prop,key) => {
-                    if(prop.path === this.props.location.pathname){
+        appRoutes.map((prop, key) => {
+            if (prop.collapse) {
+                prop.views.map((prop, key) => {
+                    if (prop.path === this.props.location.pathname) {
                         name = prop.name;
                     }
                     return null;
                 })
             } else {
-                if(prop.redirect){
-                    if(prop.path === this.props.location.pathname){
+                if (prop.redirect) {
+                    if (prop.path === this.props.location.pathname) {
                         name = prop.name;
                     }
-                }else{
-                    if(prop.path === this.props.location.pathname){
+                } else {
+                    if (prop.path === this.props.location.pathname) {
                         name = prop.name;
                     }
                 }
@@ -55,14 +78,15 @@ class Header extends Component{
         })
         return name;
     }
-    render(){
+    render() {
         return (
-            <Navbar fluid style={{position: "fixed", zIndex: 999, width: "100%"}} >
-                <Navbar.Header style={{marginLeft:20}}>
-                    <Navbar.Brand>
-                        <a href="#pablo">{this.getBrand()}</a>
+            
+            <Navbar fluid style={{ zIndex: 999, position: "fixed" }} className="navbar-container">
+                <Navbar.Header style={{ marginLeft: 20 }}>
+                    <Navbar.Brand pullLeft>
+                        <span>{this.getBrand()}</span>
                     </Navbar.Brand>
-                    <Navbar.Toggle onClick={this.mobileSidebarToggle}/>
+                    <Navbar.Toggle onClick={this.mobileSidebarToggle} />
                 </Navbar.Header>
                 <Navbar.Collapse>
                     <HeaderLinks />
