+++ src/variables/TableNonopsForm.jsx
@@ -13,9 +13,9 @@ var dataRow = {
     updatedDate: null,
     pic: "",
 }
+var check = false
 
 export default class Table extends Component {
-
     data(check) {
         $.fn.DataTable.ext.pager.numbers_length = 6;
         $.fn.dataTable.moment('DD MMMM YYYY');
@@ -58,9 +58,13 @@ export default class Table extends Component {
                 //     leftColumns: 4
                 // },
                 rowCallback(row, data, index) {
-                    $('#nonTableProgressBtn', row).click(function () {
+                    $('.nonTableProgressBtn', row).click(function () {
                         dataRow.progress = parseInt(($('select', row).val()), 10);
+                        check = true;
                     });
+                    $(row).click(function () {
+                        dataRow.id = parseInt(($(".sorting_1", row).text()), 10);
+                    })
                 },
                 columnDefs: [
                     {
@@ -68,7 +72,7 @@ export default class Table extends Component {
                         targets: 4,
                         data: 'statProgress',
                         render(data, type, row, meta) {
-                            return '<select id="nonTableProgressBtn">' +
+                            return '<select class="nonTableProgressBtn">' +
                                 '<option style = "color:#8e8e8e" value="' + row.statProgress + '" selected disabled>' + row.statProgress + '</option>' +
                                 '<option value = 10 >CLOSED</option>' +
                                 '<option value = 9 >HOLD-REJECT</option>' +
@@ -174,66 +178,61 @@ export default class Table extends Component {
 
         )
         this.$el.on('click', 'tr', function () {
-            let pos = table.row(this).index();
-            let row = table.row(pos).data();
             if ($(this).hasClass('selected')) {
                 $(this).toggleClass('selected');
-                if (dataRow.progress !== null) {
-                    dataRow.id = row.id;
-                    dataRow.updatedDate = moment();
-                    dataRow.pic = parseInt((Cookies.get('__hrni')), 10);
-
-                    var authOptions = {
-                        method: 'POST',
-                        url: 'http://0.0.0.0:8080/form/update',
-                        data: JSON.stringify(dataRow),
-                        headers: {
-                            'Content-Type': 'application/x-www-form-urlencoded'
-                        },
-                        json: true
-                    };
-                    axios(authOptions)
-                        .then(function (response) {
-                            toast.success("NonOps Form name " + row.fullName + " has been changed!!!", {
-                                position: "top-right",
-                                autoClose: 4000,
-                                hideProgressBar: true,
-                                closeOnClick: true,
-                                pauseOnHover: true,
-                                className: "notifSuccess"
-                            });
-                            dataRow.progress = null;
-                            table.ajax.reload();
-                        })
-                        .catch(function (error) {
-                            toast.error("We're sorry something wrong data hasn't been changed!!!", {
-                                position: "top-right",
-                                autoClose: 6000,
-                                hideProgressBar: true,
-                                closeOnClick: true,
-                                pauseOnHover: true,
-                                className: "notifError"
-                            });
-                        });
-                }
             }
             else {
                 table.$('tr.selected').removeClass('selected');
                 $(this).addClass('selected');
-                if (dataRow.progress !== null) {
-                    toast.error("NonOps Form name " + row.fullName + " has not been changed!!!", {
-                        position: "top-right",
-                        autoClose: 6000,
-                        hideProgressBar: true,
-                        closeOnClick: true,
-                        pauseOnHover: true,
-                        className: "notifError"
-                    });
-                }
-                dataRow.progress = null;
             }
-        });
 
+            if (check && !(isNaN(dataRow.progress))) {
+                let pos = table.row(this).index();
+                let row = table.row(pos).data();
+
+                dataRow.updatedDate = moment();
+                dataRow.pic = parseInt((Cookies.get('__hrni')), 10);
+                check = false
+                console.log("masuk", dataRow)
+
+                var authOptions = {
+                    method: 'POST',
+                    url: 'http://0.0.0.0:8080/form/update',
+                    data: JSON.stringify(dataRow),
+                    headers: {
+                        'Content-Type': 'application/x-www-form-urlencoded'
+                    },
+                    json: true
+                };
+                axios(authOptions)
+                    .then(function (response) {
+                        toast.success("NonOps Form name " + row.fullName + " has been changed!!!", {
+                            position: "top-right",
+                            autoClose: 4000,
+                            hideProgressBar: true,
+                            closeOnClick: true,
+                            pauseOnHover: true,
+                            className: "notifSuccess"
+                        });
+                        dataRow.progress = null;
+                        table.ajax.reload();
+                    })
+                    .catch(function (error) {
+                        toast.error("We're sorry something wrong data hasn't been changed!!!", {
+                            position: "top-right",
+                            autoClose: 6000,
+                            hideProgressBar: true,
+                            closeOnClick: true,
+                            pauseOnHover: true,
+                            className: "notifError"
+                        });
+                    });
+            }
+        })
+
+        setInterval(()=>{
+            table.ajax.reload();
+        },120000)
     }
 
     componentDidUpdate() {
